docker-test/dockerify/dockerify-api/requirements.txt
Django==2.2.11
djangorestframework==3.12.4
